cve_bin_tool: data_sources: Set cachedir and backupcachedir as parameter
The cachedir and backupcachedir was always on the default value for the
data sources. For the cvedb is the cachedir configurable.
With this patch the cve-bin-tool can run in different instances
with a own cache for the CVE information.

Additional the patch is a workaround for the cvedb access in line 441 purl2cpe_conn = sqlite3.connect(self.cachedir / "purl2cpe/purl2cpe.db"), which fails if cvedb has an other cachedir as the purl2cpe with the DISK_LOCATION_DEFAULT.

Signed-off-by: Maik Otto <m.otto@phytec.de>